refactored TemplateFactoryUI to UIController
